ac_mpscfifo is running.
Main issue was that the arm5 doesn't support the gcc builtin
atomic operations and I'd have to implement at least __sync_synchronize
and __atomic_exchange_4. It was easier to just change the compiler
flags. I chose to choose those for BCM2835 which has a arm1176jzf-s cpu
which is armv6zk architecture with hardware floating point.