Refactor adding platform/pc_i586.
Also, fix ac_size_t which needed to be unsigned.

I've not added testing of pc_i586 because I currently don't know how
to cause qemu to end when the application stops executing. Hopefully
it will be as easy as it was with arm where I was able to pass a
-no-reboot parameter to qemu and then reset the cpu.