test_ac_putchar runs.
The Multiboot header, mb2.S file, needs to be explicitly in the
linker command line so this it loads. So I've added it to srcFiles
rather than runtimeSrcs. Files added to runtimeSrces are added to
libruntime.a.

The -n aka --nmagic which "Do not page align data" is needed and
results in the following "Program Header:" as see with
objdump -x sadie/build-pc_x86_64/tests/test-ac_putchar/test_ac_putchar

  Program Header:
    LOAD off    0x0000000000000000 vaddr 0x0000000000000000 paddr 0x0000000000000000 align 2**21
         filesz 0x0000000000101282 memsz 0x0000000000107000 flags rwx

If there is no -n then the "Program Header:" is the following and
the multiboot header is not in the right position:

  Program Header:
    LOAD off    0x0000000000001000 vaddr 0x0000000000101000 paddr 0x0000000000101000 align 2**12
         filesz 0x0000000000000282 memsz 0x0000000000006000 flags rwx