Add ac_thread_yield and test.
ac_thread_yield is required in test-ac_thread/srcs/test.c so that the
other thread gets some cpu time and increments the counter.

For Posix ac_thread_yield just calls sched_yield.

For arch/armv6/arm1176jzf-s there was a little more work, I took the
entry and exit code from ac_exception_irq_wrapper and split it into
two routines in ac_thread.c, ac_thread_yield contains the entry code
from ac_exception_irq_wrapper and setups a stack from that the exit
code expects which was placed in ac_reschedule. So when an application
call ac_thread_yield a stack frame is setup which then branches to
ac_reschedule which will invoke thread_scheduler to find the next
ready thread to run.

Another technique would probably be needed when we split into system
and user space, but for now this looks to be working.
and this is w