Fix Posix threading.
We were not breaking out of the loop when we'd created a thread thus
multiple threads were being created ever since I changed the
ac_thread_init call in test-ac_thread/test.c from initializing 1 to
2 threads. So sometimes it would work and sometimes not just depending.
And, of course, is usually worked on my computer but not at circleci.
Good to have external testing :)

I'm now always destroying the attribute. I'm doing this because its stack
allocated its not valid after we leave the subroutine therefore it seems
reasonable to destroy it upon leaving, it it was successfully created.

Fixed a typo.