Add regs_x86.h, descriptors_x86.h, ac_architecture.h
regs_x86.h have some routines which get/set registers.

descriptors_x86.h are structures for x86_32 and x86_64 processors.

ac_architecture.h define ARCH_xxx and CPU_xxx which are used so the
regs_x86.h and descriptors_x86.h are valid for but 32 and 64 bit
x86 processors. Not fully tested but a start.