Removed most of the debug output and unused code.
We easily overrun the serial port on the real hardware, although works
fine on QEMU, and we need to add code to properly throttle the output.
Eventually we need to make it interrupt driven and message based, but
initially ac_putchar can probably just loop on the output fifo being
full.

Also, with an interrupt table the abort is not working and looks like
we get an expt_13, #GP i.e. a general protection fault. So these are
the next two problems.