Created test-cpuid_x86 from test-native_x86.
I did this so test-cpuid_x86 would run on Posix. I also expose native_x86
to posix as many of the funtions are allowed in userspace, but some are
not. In particular the test_gdt_ldt_idt fails and messing with the
segment registers probably isn't a good idea.

Further more I changed qemu-system-x86_64.runner.sh and
qemu-system-i386.runner.sh to take extra parameters. I use this when
running native and cpuid tests to enable kvm and cpu host. I did this
because there are more features available, such as x2apic using kvm.