Refactor the meson.build files.
Under some cases srcFiles always contained a reference to srcs/test.c
even if there was no test.c. This happened because srcFiles was global
and it might contain something from a previous directory.

With this change srcFiles is always local now firstSrcFiles is a global and
when generating an "executable" its required to have firstSrcFiles be
first in the srcFiles list. The is used so when building multiboot 2 images
for grub the m2b.S file can be first.