Change ac_printf to handle ac_uint/ac_sint variables.
Now %b, %d, %u and %x work for 32 or 64 bit ac_uint/ac_sint. But if
ac_sint constants are used directly in ac_printf invocations they
need to be cast. For those that desire it I've created AC_UINT,
AC_SINT and AC_INT macros which can be used to cast to their
respecitive types.