Add tests for pdte_1g_fields_array.
This works but notice that I'm not calling test_pdpte_fields_array and
instead I'm calling test_pdpte_fields with just the first test and
in test_pdpte_fields I'm not doing any of the tests.

The reason is if I do uncomment any of the tests the multiboot header
becomes corrupted and gurb complains there is no header. So we don't
even execute anything it seems we've increased the size of the code
and causing either the image to be corrupted or grub can't deal with
such a large elft image.