Move .multiboot section as first item in .text section
For some reason having KEEP(*(.multiboot)) in the .boot section caused
multiboot header code from mb2.S to be physically at the end of the elf
file, but it would be relocated to the beginning so all was well. Until
my program grew larger than 32K and the grub wouldn't find the
multiboot header.

By having it explicitly at the beginning of the .text section all is
now well!