Change some APIC tests.
Change so we expect 2 or more interrupts, this actually fails currently
because we only see 1 and we're still seeing the apci_timer_df.

Change apic_timer_df to not reset so we run more consistently. Although
it's also configured to not have an error_code even though the Intel
docs say it should.

Change divider to 128 instead of 1.

Fix test_apic_version so it ignore the support EOI broadcast bit.

Add test that the apic was initialized.