Fix issue #40 Interrupts not working correctly
Main issue was that the Double Fault was really the PIT interrupt as
Philipp suspected and my short term solution is to just handle it by
sending a EOI to the PIT using outb(0x20, 0x20). I also renamed it to
pit_isr.

Secondly, I needed to send and EOI in the apic_timer_isr also.

Another change was to not print anything in the isr as it causes the
isr to execute longer than the periodic rate of the interrupt.

Finally, cleaned up some names.