Added code to ac_init to look for MP Floating Point Structure.
This structure is part of the Intel MultiProcessing Specification, aka
MP Specificaiton. But it doesn't appear to be in memory, but there
could be a bug or I'm looking in the wrong places.

I did this because the MP configuration table has information on IO APIC's
which could possibly be used to handle issue #41, the PIT problem.