Use the clobber list to specify registers to save in isr.
This is much better, but I probably still need to save segment registers
especially FS and GS MSRs and use SWAPGS as the compiler can use them for
Thread Local Storage.

Also, the order of saving is not known so long term this is not robust
and I may need to change these to assembler unless the naked attribute
is implemented.

Bug: issue #42