Changed thread_create to current threads flags plus debug stuff in .h
Using the flags from current threads allows some control of the state
of the flags for child threads. As an example it allows a interrupts to
be disabled in the parent and also the child. Without this the child
always had interrupts enabled.

Added get_timer_schedule_isr_counter in .h

Added print_ready_list for debugging.