Changed ac_receptor_signal.
Changed to void return and added yield_if_waiting so caller can decide
whether or not to yield if there is a waiting thread when we signal.
This gives the caller more precise control of their liveliness.

ac_thread_wait and _signal are "working". There are two implementations
of signal one which uses a loop and the other which disables interrupts.
Right now I've enabled the loop.