Change thread_x86 to use apic TSC-Deadline for timer interrupts.
TSC-Deadline is much more accurate and allows the "slice" to be any
value because delays can be arbitrary in length. The one-shot mode has
a much more limited duration and precision.

Add test_thread_wait which is from tests/test-ac_thread but having it
here allows for the addition of X86 speicfic tests.

Because we're using TSC-Deadline needed to change run-test-thread_x86 to
use enable-kvm.