Add a priority queue for managing thread waiting.
Now multiple threads can wait simultaneously. I've not measured the
efficiency but it seems to be working. Insertion seems to be fairly
quick but removal is taking more code than I'd like. So something
to look into someday.

Also, because of the number of threads and I haven't implemented
ac_free I had to increase the memory size to 500K.