Have each undefined interrupt be unique.
This allows you to know which interrupt fired, I'm hoping it will be
useful for debugging ISR's. There is probably a better way to do this
but this seems to be working for now.

Also added AC_CAT which allows concatenation of two tokens and can be
used as token. In this change look at the INTR_UNDEFINED macro defined
in interrupts_x86.c