Fix failure in testing ac_time.
I needed to have ac_tsc_init at least initialize the freq and divider.
To do this I needed to disable the MCR instructions which are causing
undefined exceptions.

I also needed the ac_sprintf calls in ac_time to use %ld to since
ARM is 32 bit processor and the variables were 64 bit.