Fix issue #55 and refactoring
I've refactor ac_mpscfifo to not need the stub to be passed. But now
there is an ac_mpscfifo_early_init which must be called. For Posix
its a constructor and pc_x86_64 its called from ac_init.

Also, changed AcMsgPool to not be a pointer as so it will be easier to
expose its opaque struct in the future and be more consistent.