Fix issue #56.
Primary bug is that remove_zombies needed to be called, this can be
done either by waiting long enough so the idle thread gets executed
or we call it directly, for now we'll call it directly.

Also fixed mptt to return the dispatcher and deinit the mpscfifo, this
recovers the dispatcher, mpscfifo's and msgs. We still don't have a way
to return receptors, issue #61.