Use __atomic_thread_fence instead of __atomic_fetch_add
Since this MsgTsc information is logging information no memory is
concurrently modified so the atomic_fetch_add isn't necessary. But
to be sure the cache is flushed I've added a __atomic_thread_fence at
the end of the routine.

In reality I think I need to add the __atomic_thread_fence when the
thread terminates to be sure cache's are flushed. I've added issue #62
to track this.