Change AC_TEST and AC_TEST_EM to output "ERROR:+:"
When running tests with qemu the error exit of the application running
within qemu isn't available. So instead the qemu-*.runner.sh captures
the output and looks for a pattern. I've changed the pattern to
"ERROR:+:" as an unlikely output from other apps and thus identifying
errors from my tests.