Change ac_printf/*/test.c to use the ERROR:+: pattern for errors
Previously it was outputting FAILURE and errors weren't being detected
by qemu-*-runner.sh. This should cause a failure to be detected in a test
of print_ff_format_proc on Platform=VersatilePB. There isn't an error in
Posix or pc_x86_64 so I'm guessing the problem is related to processing
var args.