Changed ac_printf_format_proc to return args consumed
This is necessary because of the difference between how x86_64 and
ARM gcc compilers handled __builtin_va_arg in subroutines. In the
x86_64 a consumed value updates the list argument passed so its not
necessary to return the count.

But in the ARM its not, so I changed the invocation of a
ac_printf_format_proce in ac_printf to pass a copy of the args and
have the caller return the number of ac_uint's consumed. This makes
it less portable but its the best I can do and the test I have in
test.c does catch the difference between ARM and x86_64.