AcBuff and AcMpscFifo are passing there tests.
One thing I'm not happy with is that AcBuff_alloc is returning an array
of AcBuffs and maybe it should return an array of AcBuff* i.e. pointers.
This way AcBuff_get_nth would not be needed.

Of course if I really wanted that then I could call AcBuff_alloc with
a count of 1. So the current implementation is better in the sense the
caller can decide to get pointers to individual elements by calling it
with the count of 1 or get a bunch with a count > 1.

I could also provide a helper routine which would create or init an array
of pointers. So lets see how it goes, in the end I don't thing user code
will be dealing the the array's directly they'll be using pools which
can handle it appropriately.