WIP: Add AcMpscFifo_deinit_full.
Added AcMpscFifo.count and now AcMpscFifo_deinit assumes a fifo created
with AcMem_alloc with a count == 1, i.e. the stub. This type of fifo
is typically used for things like communication between components.

Added AcMpscFifo_deinit_full which assumes a fifo that was allocated
a created with AcMem_alloc with the count parameter > 1. This type of
fifo is typically used for a pool.