Change ac_comp_mgr and ac_dispatcher to use AcMessage and AcMessagePool
This necessitated changing ac_bits to not use expression statements as
they are not constants and can't be used in switch statements which was
also the reason for creating the OPERATION macro.

I had to disable test_ac_msg_pool_multiple_threads but that will be going
away now that we have AcMessagePool.

I also disabled tests/test-ac_dispatcher as its not yet converted to use
AcMessage/AcMessagePool.