Add trying two different interface names.
Made AcCompIpv4LinkLayer.ifname an array and change the INIT_CMD to try
both. The names are "eno1" for Arch Linux and "eth0" for Ubuntu which
is what circleci is using.

So I was able to use sudo in circleci and with the right interface
name I'm able to open it.